content/requirements.txt
Flask==2.0.1
Flask-API==3.0.post1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
flask-restplus==0.13.0
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
glog==0.3.1
numpy==1.18.2
pika==1.1.0
PyJWT==1.7.1